Unify hex codec helpers in @interchange/types
Five copies of hexEncode and four copies of hexDecode were spread
across hub, sidecar, and test files — each independently re-derived
and one of them validation-light. Collapse them into a single pair
exported from @interchange/types, where the existing address parser
already lives. The validated decoder uses the sidecar's "odd-length"
and "non-hex characters" error wording so the existing test that
asserts on that message keeps passing.

The hub's previous hexDecode used different error text but its
errors were always swallowed at the call site, so unifying on the
sidecar's wording is observable only through the now-shared
implementation.
