+++ packages/hub/src/ws/sidecar-handler.test.ts
@@ -4,22 +4,8 @@ import {
   generateKeyPair,
   importPrivateKeyBytes,
 } from "@interchange/crypto-node";
-import { parseAgentAddress } from "@interchange/types";
+import { hexDecode, hexEncode, parseAgentAddress } from "@interchange/types";
 import { createSidecarRouter, type WsHandle } from "./sidecar-handler";
-
-function hexEncode(bytes: Uint8Array): string {
-  return Array.from(bytes)
-    .map((b) => b.toString(16).padStart(2, "0"))
-    .join("");
-}
-
-function hexDecode(hex: string): Uint8Array {
-  const bytes = new Uint8Array(hex.length / 2);
-  for (let i = 0; i < bytes.length; i++) {
-    bytes[i] = parseInt(hex.substring(i * 2, i * 2 + 2), 16);
-  }
-  return bytes;
-}
 
 function signChallenge(
   nonce: string,
